# Printer3D-backend/src/spec/printer3d-openapi.yaml
...
openapi: 3.0.0
info:
  title: 3D Printer Queue Management API
  description: >
    # Summary

    ### This is the OpenApi specification active and matching the latest published backend application version seen on the Heroku Production servers.
    This API defines all the client REST endpoints public and available to other systems. The API definition follows the OpenApi 3.0 specification
    and there are provisioning for automatic code generation for this API clients.

    # System Description

    The system is composed of 4 subsystems interconnected by interchange lines. The first system represents the *external user demand*. This system
    has a catalog of items and generates a queue of Requests. This is connected to the second subsystem that is the *Storage* system. The storage
    contains the elements available to other systems or the elements to generate Parts and Models. This second system connects with the third
    subsystem called *Production* that is the responsible to calculate and control the elements to be produced and the tool to build them. The input
    are the demands from Storage to keep the stocks or to fulfill the customer Requests. And finally this reports to the four subsystem to buy
    materials and tools to maintain the Production operative and at the right levels.


    Not all the systems have an application implementation. There are two main applications. One called **frontend** to deal with the Graphical User
...


...

    # API Description

...


...

...

...

...

...

...

...
servers:
...
paths:
...
  /api/v1/inventory/parts:
    post:
      tags:
        - Inventory
      summary: Create a new Part from the data on the request.
      description: >
        The Printer3D user interface should have requested the Part contents to the user. This endpoint should validate all the fields against the
        validation requirements and create a new record on the Inventory repository for the new Part.
      operationId: newPart
      requestBody:
        description: >
          Contains the Part fields to be used to create a new Part record at the repository.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Part'
      responses:
        201:
          description: >
            The new Part as stored at the Inventory repository.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Part'
        409:
          description: >
            There is a conflict on the repository related to this Part. There can be some causes.
            * PART_ALREADY_EXISTS - The new Part identifier is already used by a repository Part. The record cannot be created.
            * PART_REPOSITORY_CONFLICT - There is a constraint error detected. Probably that the key fields of the Part atch the contents of an
            already existing part at the repository.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Part'
    patch:
      tags:
        - Inventory
      summary: Update a Part with some of the fields on the request.
      description: >
        The Printer3D user interface now has a feature tochange some of the Part fields. Thew are mostly related with the stock values and the cost
        data. When the PATCH operation is sent to the backend all the fileds are populated but this endpoint will only get some of them and persist
        their new values on the repository.
      operationId: updatePart
      requestBody:
        description: >
          Contains the Part fields to be used to update the Part record at the repository.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Part'
      responses:
        201:
          description: >
            The new Part as stored at the Inventory repository.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Part'
    get:
      tags:
        - Inventory
      summary: Get the list of Parts persisted at the Inventory repository.
      description: >
        Get the complete list of <b>Parts</b> persisted at the Inventory repository. If the active filter is active retrieve only the active Parts.
      operationId: getParts
      parameters:
        - in: query
          name: activesOnly
          required: false
          style: form
          explode: false
          description: >
            Allows the selection or filtering for not active Parts. By default all active parts are retrieved. If false all parts are included on the list. If tru only active parts will be returned by the filter.
          schema:
            type: boolean
            default: false
      responses:
        200:
          description: >
            The list of Parts persisted at the Inventory repository.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartList'
  /api/v1/inventory/parts/group:
    patch:
      tags:
        - Inventory
      summary: Update a group of Parts that have a common label.
      description: >
        Usually there a re a set of parts that share most of the field values changing only the finishings (material and color). Because the
        frontend will group them into sets and there is no purpose on having different values for same parts this endpoint will take care of
        changing the et values to the whole set of parts that have the same Label.
      operationId: updateGroupPart
      requestBody:
        description: >
          Contains the Part fields to be used to update the Part record at the repository.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupPartRequest'
      responses:
        201:
          description: >
            The new Part as stored at the Inventory repository.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CounterResponse'

  # - M O D E L S
...
  /api/v2/production/requests:
    post:
      tags:
        - Production
      summary: Creates a new Request instance at the persistence repository.
      description: >
        This is an entity used by the customers to report the items they want to buy. The **Request** contains the list of
            specific Part models along with the material and color that are required by the customer. Also Requests can contain Models that are packaged
            sets of Parts that are sell together. Requests are processed by order of arrival when the list of current open requests is solicited by the
            frontend. The system will follow the first come/first served pattern. Then Part build jobs generated by Requests have a higher priority than
            other jobs generated by inventory leveling.
            A **Request** can be on different states. It starts on the state **OPEN** when the request is created. When the frontend requests the list of
            open
            buy requests then it is compared against the current stock levels. If all the items on the request are available then it moves to the state
            **COMPLETED**. If the user at the frontend delivers the request to the customer it can then close is and remove the items from the inventory. At
            this point the request moved to the final state of **CLOSED**.
      operationId: newRequest
      requestBody:
        description: >
          Contains the **Request** field values filled at the UI.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestV2'
      responses:
        201:
          description: >
            The new Request as transformed from the persistence repository.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestV2'
        400:
          description: >
            If the request to create a new requst has not all the required fields or they do not pass the validations then we reject the request
            with a bad format request exception. The response message should explain the cause for the rejection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DimensinfinError'
    patch:
      tags:
        - Production
      summary: Updates an Open Request instance at the persistence repository.
      description: >
        An open Request can be edited at the UI. The user can decide to reduce the number of Parts initially requested or he can add or remove
        Parts from the Request. Once the Request is completed and closed it is not editable..
      operationId: updateRequest
      requestBody:
        description: >
          Contains the **Request** field values filled at the UI.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestV2'
      responses:
        200:
          description: >
            The new Request as transformed from the persistence repository.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestV2'
    get:
      tags:
        - Production
      summary: Get the list of Requests persisted at the repository.
      description: >
        Get the complete list of **Requests** persisted at the Production repository and that are on an Open state.
        This method is compatible with the implementation of Requests V1 until the repository V1 is deprecated and defiled.
        The process now extends to support Models that are packs of Parts. The new extension expands the list from the Model BOM to check agains
        the current stock of Parts to check for run out of instances. Complete requests continue to be marked COMPLETED.
        During the processing each of the requests in FIFO sequence is processed against the Part stocks. If there are stocks for all the Parts in
        the request the state changes to COMPLETED during this request. The Part stocks are updated for each Request processed.
      operationId: getOpenRequests
      responses:
        200:
          description: >
            The list of Requests in the Open state as transformed from the repository stored copy.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RequestV2'
    delete:
      tags:
        - Production
      summary: Delete the selected Request form the repository.
      description: >
        Requests have no other relations with repository records so they can be deleted safely. If the user select to remove a Request because it
        should be modified of because no longer required the record ban be deleted without impact. Only OPEN Requests can be deleted. CLOSED are no
        longer available for update or deletion.
      operationId: deleteRequest
      parameters:
        - in: path
          name: requestId
          required: true
          style: simple
          explode: false
          description: The unique id (uuid) of the Request we should delete.
          schema:
            type: string
            format: uuid
          example:
            4e7001ee-6bf5-40b4-9c15-61802e4c59ea
      responses:
        200:
          description: >
            Delete the selected request from the repository. The endpoint received the unique request identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CounterResponse'
  /api/v2/production/requests/{requestId}/close:
    put:
      tags:
        - Production
      summary: Closes the state of a selected Request.
      description: >
        When the Request is closed is the moment where the Parts that compose the request are subtracted from the Parts inventory. The close should
        be able to close requests on the old repository and also on the new repository. So the identifier should be searched on both repositories
        to locate the right instance to be closed.
      operationId: closeRequest
      parameters:
        - in: path
          name: requestId
          required: true
          style: simple
          explode: false
          description: The unique id (uuid) of the Request we are closing.
          schema:
            type: string
            format: uuid
          example:
            4e7001ee-6bf5-40b4-9c15-61802e4c59ea
      responses:
        200:
          description: >
            Returns the current state of the Request on the repository.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestV2'
        404:
          description: >
            If the request searched by the request id is not found on neither of the v1 or v2 repositories then we should raise this exception.
            This has to be reported to the system administrator because there can show a data corruption problem.
            Other cause can be that the Request has been closed before on another session and that frontend data is stale.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DimensinfinError'

  # - J O B S
...
components:
  schemas:
    Part:
      type: object
      required:
        - id
        - label
        - colorCode
        - buildTime
        - cost
        - price
        - stockLevel
      properties:
        id:
          type: string
          format: uuid
          pattern: "[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}"
          description: >
            This is a frontend generated unique uuid assigned when the Part is created. It cannot be changed on the Part lifecycle and it is used
            as the single unique identifier to deal with the Part or its derivatives.
          example:
            4e7001ee-6bf5-40b4-9c15-61802e4c59ea
        label:
          type: string
          format: string
          minLength: 3
          maxLength: 50
          description: >
            It is a human readable label to identify the product. It is a product code but that it usually it is not unique since the same part
            design can be used with different finishings or colours. Cannot be changed for the lifetime of the Part and it is keyed by the user at
            the front end interface.
          example:
            Covid-19 Key
        description:
          type: string
          format: string
          maxLength: 500
          description: >
            A user description of the part. It is useful but it is not mandatory.
          example:
            'This is a key to be used to isolate contact with surfaces and buttons. Use it to open doors and push buttons.'
        material:
          type: string
          format: string
          minLength: 1
          maxLength: 16
          description: >
            It is the acronym or code for the type of plastic. There are different types of plastics for different uses and some of them require an
            specialized machine or a particular machine configuration.
          example:
            PLA
        color:
          type: string
          format: string
          minLength: 1
          maxLength: 32
          description: >
            The color name for the plastic on the coil. A coil should have this field as mandatory
          example:
            VERDE-T
        weight:
          type: number
          format: int32
          default: 1
          minimum: 1
          maximum: 400
          description: >
            This is the weigth in grames consumed by the Machine to build one instance of this Part.
          example:
            12
        buildTime:
          type: number
          format: int32
          default: 5
          description: >
            This field stores the time in minutes that lasts the 3d printer job to finish.
          example:
            15
        cost:
          type: number
          format: float
          description: >
            Cost estimated value in Euros with two decimals.
          example:
            0.85
        price:
          type: number
          format: float
          description: >
            The requested catalog price for this Part. This price includes IVA so on bills real price should be recalculated.
          example:
            4.00
        stockLevel:
          type: number
          format: int32
          default: 1
          description: >
            This is the value set as the stock level desired to have on the storage.
          example:
            3
        stockAvailable:
          type: number
          format: int32
          default: 0
          description: >
            This is the current stock value of the number of pieces available to sell.
          example:
            4
        imagePath:
          type: string
          format: url
          description: >
            The URL for an external image stored on a public repository or Google drive. This image will be used when showing the catalog.
          example:
            'https://ibb.co/3dGbsRh'
        modelPath:
          type: string
          format: string
          description: >
            The file name os the model to be used on the printer to be used on the printer job.
          example:
            'pieza3.STL'
        active:
          type: boolean
          format: boolean
          default: true
          description: >
            This flag is used to mark out disabled parts. Parts that are obsolete and no longer should be build are not deleted but just disabled
            with this flag. Disable status can be reverted editing the part.
          example:
            true
    PartList:
      type: object
      required:
        - count
        - parts
      properties:
        count:
          type: number
          format: int32
          description: >
            The number of parts that follow on the response.
          example:
            8
        parts:
          type: array
          items:
            $ref: '#/components/schemas/Part'
    UpdateGroupPartRequest:
      type: object
      properties:
        label:
          type: string
          format: string
          minLength: 3
          maxLength: 50
          description: >
            It is a human readable label to identify the product. It is a product code but that it usually it is not unique since the same part
            design can be used with different finishings or colours. Cannot be changed for the lifetime of the Part and it is keyed by the user at
            the front end interface.
          example:
            Covid-19 Key
        description:
          type: string
          format: string
          maxLength: 500
          description: >
            A user description of the part. It is useful but it is not mandatory.
          example:
            'This is a key to be used to isolate contact with surfaces and buttons. Use it to open doors and push buttons.'
        weight:
          type: number
          format: int32
          default: 1
          minimum: 1
          maximum: 400
          description: >
            This is the weigth in grames consumed by the Machine to build one instance of this Part.
          example:
            12
        buildTime:
          type: number
          format: int32
          default: 5
          description: >
            This field stores the time in minutes that lasts the 3d printer job to finish.
          example:
            15
        imagePath:
          type: string
          format: url
          description: >
            The URL for an external image stored on a public repository or Google drive. This image will be used when showing the catalog.
          example:
            'https://ibb.co/3dGbsRh'
        modelPath:
          type: string
          format: string
          description: >
            The file name os the model to be used on the printer to be used on the printer job.
          example:
            'pieza3.STL'
...
    RequestV2:
      type: object
      description: >
        This is the data structure that is received/sent over the HTTP interface. When the user creates or changes the Request the service will
        receive a copy of this structure and then return an updated and persisted copy.
        Version 2 requests are able to contain any type of element and not only Parts. The contents definitions then changes to allow this different
        way to store the request contents information. Persisted records of version 2 are not compatible with persisted records of version 1 so
        records of version 1 are expected to be closed and not to be returned on the request.
        If anyway the system still has V1 records they are converted to the V2 format (part identifiers are wrapped again) and then processed and
        reported to the frontend.
      required:
        - id
        - requestDate
        - label
        - contents
        - state
      properties:
        id:
          type: string
          format: uuid
          pattern: "[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}"
          description: >
            The unique identifier for a single Request. Once a request is closed cannot be edited so the user will need to create a new instance.
            This field is used for primary keying the requests.
          example:
            009ab011-03ad-4e84-9a88-25708d1cfd64
        requestDate:
          type: string
          format: date-time
          description: >
            The date and time when the Request is submitted to the backend. Used for request ordering on the FIFO request list.
          example:
            2020-06-01T22:05:00Z
        label:
          type: string
          format: string
          minLength: 3
          maxLength: 50
          description: >
            The user readable label to be assigned to a single Machine. Can be used instead of the unique identifier UUID.
          example:
            'Machine A'
        contents:
          type: array
          items:
            $ref: '#/components/schemas/RequestItem'
        state:
          type: string
          enum:
            - OPEN
            - COMPLETED
            - CLOSED
          default: OPEN
          description: >
            Requests that should be sent to the customer but that still have missing Parts should stay at the OPEN state until the Request is
            completed and it is served. When the transaction completes then the requests moves to the CLOSED state and then cannot be changed.
            State descriptions:
            * **OPEN** - The Requests are created on this state. This is the value stored at the database while the Request is not closed. Open
            requests are listed to the frontend and at that moment they can change to COMPLETED if there are enough Parts.
            * **COMPLETED** - This is the state reflected when the Request is listed to the frontend. This state is not stored at the repository.
            Means that the Request can be served to the customer and then closed.
            * **CLOSED** - The Request has been sent to the customer and the Parts removed from the inventory stock. This is the final state for
            the Request.
...
    CounterResponse:
      type: object
      properties:
        records:
          type: number
          format: int32
          description: >
            The number of records processed by the previous operation. Applies to deletions and also to updates.
          example:
            3
...